Move NavMenu index out of NavMenu directory and rename to NavMenu
