+++ client/public/index.html
@@ -11,7 +11,6 @@
     <link
       href="https://fonts.googleapis.com/css2?family=Cormorant+Garamond:wght@400;700&family=Fira+Mono&family=Montserrat:wght@300;400;700&family=Pinyon+Script&display=swap"
       rel="stylesheet">
-      <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
     <title>Your Memories, Your Journey, Beautifully Preserved | WEDLOC</title>
   </head>
   <body>
