--- client/src/components/NavMenu/index.css
@@ -1,61 +1,0 @@
-/* NAV STYLES */
-
-.menu-trigger {
-    position: absolute;
-    top: 1vh;
-    right: 10px;
-    height: 5vh;
-    width: 3vw;
-    cursor: pointer;
-}
-
-.menu-trigger img {
-    position: absolute;
-    top: 10px;
-    right: 10px;
-    height: 5vh;
-    width: 3vw;
-    border-radius: 30px;
-    cursor: pointer;
-}
-
-.dropdown-menu {
-    position: absolute;
-    top: 90px;
-    right: 10px;
-    border-radius: 30px;
-    padding: 10px 20px;
-    width: 11vw;
-    border: solid rgb(169 169 169 / 15%);
-    font-size: large;
-    background: rosybrown;
-}
-
-.dropdown-menu::before {
-    content: '';
-    position: absolute;
-    top: -5px;
-    right: 10px;
-    height: 20px;
-    width: 20px;
-    transform: rotate(45deg);
-}
-
-.dropdown-menu.active {
-    opacity: 1;
-    visibility: visible;
-    transform: translateY(0);
-    transition: 500ms ease;
-}
-
-.dropdown-menu.inactive {
-    opacity: 0;
-    visibility: hidden;
-    transform: translateY(-20px);
-    transition: 500ms ease;
-}
-
-.liNav {
-    list-style: none;
-    padding: 10px;
-}
